+++ packages/spec/.eslintrc.common-ts.js
@@ -1,0 +1,7 @@
+const { getESLintConfig } = require('./src');
+
+module.exports = getESLintConfig('common-ts', {
+  env: {
+    jest: true,
+  },
+});