+++ packages/spec/examples/common-ts/index.ts
@@ -1,0 +1,7 @@
+interface AppDeveloper {
+  name: string;
+  id: number;
+}
+
+const bar = [1, 2];
+const one: AppDeveloper = { name: 'hello', id: 1 };