+++ packages/spec/src/commitlint/common.js
@@ -1,0 +1,4 @@
+// commitlint config for common project
+module.exports = {
+  extends: 'ali',
+};