+++ packages/spec/src/prettier/common.js
@@ -1,0 +1,8 @@
+// prettier config for common project
+module.exports = {
+  printWidth: 120,
+  tabWidth: 2,
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'all',
+};