+++ packages/spec/CHANGELOG.md
@@ -1,4 +1,8 @@
 # changelog
 
+## 1.2.0
+- ESlint config support `common` and `common-ts` rules.
+- Commitlint, prettier, stylelint config support `common` rules.
+
 ## 1.1.0
 - Support `vue` and `vue-ts` project.